test int32_t

+++ sdk/CPlayer_ItemServices.h
@@ -54,7 +54,7 @@ public:
 	SCHEMA_FIELD(int32_t, CEconItemView, m_OriginalOwnerXuidLow);
 	SCHEMA_FIELD(int32_t, CEconItemView, m_OriginalOwnerXuidHigh);
 	SCHEMA_FIELD(uint32_t, CEconItemView, m_iAccountID);
-	SCHEMA_FIELD(uint16_t, CEconItemView, m_iItemDefinitionIndex);
+	SCHEMA_FIELD(int32_t, CEconItemView, m_iItemDefinitionIndex);
 	SCHEMA_FIELD(bool, CEconItemView, m_bInitialized);
 	SCHEMA_FIELD(char[32], CEconItemView, m_szCustomName);
 };
@@ -76,7 +76,7 @@ public:
 	SCHEMA_FIELD(uint64_t, CEconEntity, m_OriginalOwnerXuidLow);
 	SCHEMA_FIELD(uint32_t, CBaseEntity, m_nSubclassID);
 	SCHEMA_FIELD(int32_t, CBaseEntity, m_iOldOwnerClass);
-	SCHEMA_FIELD(uint16_t, CEconEntity, m_iItemDefinitionIndex);
+	SCHEMA_FIELD(int32_t, CEconEntity, m_iItemDefinitionIndex);
 };
 
 class CPlayer_WeaponServices : public CPlayerPawnComponent
